+++ eas-20210701/alibabacloud_eas20210701/models.py
@@ -348,10 +348,18 @@ class Resource(TeaModel):
         self,
         cluster_id: str = None,
         cpu_count: int = None,
+        cpu_used: int = None,
         create_time: str = None,
         extra_data: Dict[str, Any] = None,
+        features: List[str] = None,
         gpu_count: int = None,
+        gpu_used: float = None,
         instance_count: int = None,
+        instance_max_allocatable_cpu: int = None,
+        instance_max_allocatable_gpu: float = None,
+        instance_max_allocatable_memory: int = None,
+        memory: int = None,
+        memory_used: int = None,
         message: str = None,
         post_paid_instance_count: int = None,
         pre_paid_instance_count: int = None,
@@ -364,10 +372,18 @@ class Resource(TeaModel):
     ):
         self.cluster_id = cluster_id
         self.cpu_count = cpu_count
+        self.cpu_used = cpu_used
         self.create_time = create_time
         self.extra_data = extra_data
+        self.features = features
         self.gpu_count = gpu_count
+        self.gpu_used = gpu_used
         self.instance_count = instance_count
+        self.instance_max_allocatable_cpu = instance_max_allocatable_cpu
+        self.instance_max_allocatable_gpu = instance_max_allocatable_gpu
+        self.instance_max_allocatable_memory = instance_max_allocatable_memory
+        self.memory = memory
+        self.memory_used = memory_used
         self.message = message
         self.post_paid_instance_count = post_paid_instance_count
         self.pre_paid_instance_count = pre_paid_instance_count
@@ -391,14 +407,30 @@ class Resource(TeaModel):
             result['ClusterId'] = self.cluster_id
         if self.cpu_count is not None:
             result['CpuCount'] = self.cpu_count
+        if self.cpu_used is not None:
+            result['CpuUsed'] = self.cpu_used
         if self.create_time is not None:
             result['CreateTime'] = self.create_time
         if self.extra_data is not None:
             result['ExtraData'] = self.extra_data
+        if self.features is not None:
+            result['Features'] = self.features
         if self.gpu_count is not None:
             result['GpuCount'] = self.gpu_count
+        if self.gpu_used is not None:
+            result['GpuUsed'] = self.gpu_used
         if self.instance_count is not None:
             result['InstanceCount'] = self.instance_count
+        if self.instance_max_allocatable_cpu is not None:
+            result['InstanceMaxAllocatableCPU'] = self.instance_max_allocatable_cpu
+        if self.instance_max_allocatable_gpu is not None:
+            result['InstanceMaxAllocatableGPU'] = self.instance_max_allocatable_gpu
+        if self.instance_max_allocatable_memory is not None:
+            result['InstanceMaxAllocatableMemory'] = self.instance_max_allocatable_memory
+        if self.memory is not None:
+            result['Memory'] = self.memory
+        if self.memory_used is not None:
+            result['MemoryUsed'] = self.memory_used
         if self.message is not None:
             result['Message'] = self.message
         if self.post_paid_instance_count is not None:
@@ -425,14 +457,30 @@ class Resource(TeaModel):
             self.cluster_id = m.get('ClusterId')
         if m.get('CpuCount') is not None:
             self.cpu_count = m.get('CpuCount')
+        if m.get('CpuUsed') is not None:
+            self.cpu_used = m.get('CpuUsed')
         if m.get('CreateTime') is not None:
             self.create_time = m.get('CreateTime')
         if m.get('ExtraData') is not None:
             self.extra_data = m.get('ExtraData')
+        if m.get('Features') is not None:
+            self.features = m.get('Features')
         if m.get('GpuCount') is not None:
             self.gpu_count = m.get('GpuCount')
+        if m.get('GpuUsed') is not None:
+            self.gpu_used = m.get('GpuUsed')
         if m.get('InstanceCount') is not None:
             self.instance_count = m.get('InstanceCount')
+        if m.get('InstanceMaxAllocatableCPU') is not None:
+            self.instance_max_allocatable_cpu = m.get('InstanceMaxAllocatableCPU')
+        if m.get('InstanceMaxAllocatableGPU') is not None:
+            self.instance_max_allocatable_gpu = m.get('InstanceMaxAllocatableGPU')
+        if m.get('InstanceMaxAllocatableMemory') is not None:
+            self.instance_max_allocatable_memory = m.get('InstanceMaxAllocatableMemory')
+        if m.get('Memory') is not None:
+            self.memory = m.get('Memory')
+        if m.get('MemoryUsed') is not None:
+            self.memory_used = m.get('MemoryUsed')
         if m.get('Message') is not None:
             self.message = m.get('Message')
         if m.get('PostPaidInstanceCount') is not None:
@@ -753,6 +801,45 @@ class ResourceInstanceWorker(TeaModel):
         return self
 
 
+class ServiceInstanceCountInResource(TeaModel):
+    def __init__(
+        self,
+        dedicated: int = None,
+        public: int = None,
+        quota: int = None,
+    ):
+        self.dedicated = dedicated
+        self.public = public
+        self.quota = quota
+
+    def validate(self):
+        pass
+
+    def to_map(self):
+        _map = super().to_map()
+        if _map is not None:
+            return _map
+
+        result = dict()
+        if self.dedicated is not None:
+            result['Dedicated'] = self.dedicated
+        if self.public is not None:
+            result['Public'] = self.public
+        if self.quota is not None:
+            result['Quota'] = self.quota
+        return result
+
+    def from_map(self, m: dict = None):
+        m = m or dict()
+        if m.get('Dedicated') is not None:
+            self.dedicated = m.get('Dedicated')
+        if m.get('Public') is not None:
+            self.public = m.get('Public')
+        if m.get('Quota') is not None:
+            self.quota = m.get('Quota')
+        return self
+
+
 class ServiceLabels(TeaModel):
     def __init__(
         self,
@@ -794,14 +881,19 @@ class Service(TeaModel):
         app_spec_name: str = None,
         app_type: str = None,
         app_version: str = None,
+        autoscaler_enabled: bool = None,
         caller_uid: str = None,
         cpu: int = None,
         create_time: str = None,
+        cronscaler_enabled: bool = None,
         current_version: int = None,
         extra_data: str = None,
+        gpucore_percentage: int = None,
+        gpumemory: int = None,
         gateway: str = None,
         gpu: int = None,
         image: str = None,
+        instance_count_in_resource: ServiceInstanceCountInResource = None,
         internet_endpoint: str = None,
         intranet_endpoint: str = None,
         labels: List[ServiceLabels] = None,
@@ -817,6 +909,7 @@ class Service(TeaModel):
         request_id: str = None,
         resource: str = None,
         resource_alias: str = None,
+        resource_burstable: bool = None,
         role: str = None,
         role_attrs: str = None,
         running_instance: int = None,
@@ -841,14 +934,19 @@ class Service(TeaModel):
         self.app_spec_name = app_spec_name
         self.app_type = app_type
         self.app_version = app_version
+        self.autoscaler_enabled = autoscaler_enabled
         self.caller_uid = caller_uid
         self.cpu = cpu
         self.create_time = create_time
+        self.cronscaler_enabled = cronscaler_enabled
         self.current_version = current_version
         self.extra_data = extra_data
+        self.gpucore_percentage = gpucore_percentage
+        self.gpumemory = gpumemory
         self.gateway = gateway
         self.gpu = gpu
         self.image = image
+        self.instance_count_in_resource = instance_count_in_resource
         self.internet_endpoint = internet_endpoint
         self.intranet_endpoint = intranet_endpoint
         self.labels = labels
@@ -864,6 +962,7 @@ class Service(TeaModel):
         self.request_id = request_id
         self.resource = resource
         self.resource_alias = resource_alias
+        self.resource_burstable = resource_burstable
         self.role = role
         self.role_attrs = role_attrs
         self.running_instance = running_instance
@@ -884,6 +983,8 @@ class Service(TeaModel):
         self.workspace_id = workspace_id
 
     def validate(self):
+        if self.instance_count_in_resource:
+            self.instance_count_in_resource.validate()
         if self.labels:
             for k in self.labels:
                 if k:
@@ -905,22 +1006,32 @@ class Service(TeaModel):
             result['AppType'] = self.app_type
         if self.app_version is not None:
             result['AppVersion'] = self.app_version
+        if self.autoscaler_enabled is not None:
+            result['AutoscalerEnabled'] = self.autoscaler_enabled
         if self.caller_uid is not None:
             result['CallerUid'] = self.caller_uid
         if self.cpu is not None:
             result['Cpu'] = self.cpu
         if self.create_time is not None:
             result['CreateTime'] = self.create_time
+        if self.cronscaler_enabled is not None:
+            result['CronscalerEnabled'] = self.cronscaler_enabled
         if self.current_version is not None:
             result['CurrentVersion'] = self.current_version
         if self.extra_data is not None:
             result['ExtraData'] = self.extra_data
+        if self.gpucore_percentage is not None:
+            result['GPUCorePercentage'] = self.gpucore_percentage
+        if self.gpumemory is not None:
+            result['GPUMemory'] = self.gpumemory
         if self.gateway is not None:
             result['Gateway'] = self.gateway
         if self.gpu is not None:
             result['Gpu'] = self.gpu
         if self.image is not None:
             result['Image'] = self.image
+        if self.instance_count_in_resource is not None:
+            result['InstanceCountInResource'] = self.instance_count_in_resource.to_map()
         if self.internet_endpoint is not None:
             result['InternetEndpoint'] = self.internet_endpoint
         if self.intranet_endpoint is not None:
@@ -953,6 +1064,8 @@ class Service(TeaModel):
             result['Resource'] = self.resource
         if self.resource_alias is not None:
             result['ResourceAlias'] = self.resource_alias
+        if self.resource_burstable is not None:
+            result['ResourceBurstable'] = self.resource_burstable
         if self.role is not None:
             result['Role'] = self.role
         if self.role_attrs is not None:
@@ -1003,22 +1116,33 @@ class Service(TeaModel):
             self.app_type = m.get('AppType')
         if m.get('AppVersion') is not None:
             self.app_version = m.get('AppVersion')
+        if m.get('AutoscalerEnabled') is not None:
+            self.autoscaler_enabled = m.get('AutoscalerEnabled')
         if m.get('CallerUid') is not None:
             self.caller_uid = m.get('CallerUid')
         if m.get('Cpu') is not None:
             self.cpu = m.get('Cpu')
         if m.get('CreateTime') is not None:
             self.create_time = m.get('CreateTime')
+        if m.get('CronscalerEnabled') is not None:
+            self.cronscaler_enabled = m.get('CronscalerEnabled')
         if m.get('CurrentVersion') is not None:
             self.current_version = m.get('CurrentVersion')
         if m.get('ExtraData') is not None:
             self.extra_data = m.get('ExtraData')
+        if m.get('GPUCorePercentage') is not None:
+            self.gpucore_percentage = m.get('GPUCorePercentage')
+        if m.get('GPUMemory') is not None:
+            self.gpumemory = m.get('GPUMemory')
         if m.get('Gateway') is not None:
             self.gateway = m.get('Gateway')
         if m.get('Gpu') is not None:
             self.gpu = m.get('Gpu')
         if m.get('Image') is not None:
             self.image = m.get('Image')
+        if m.get('InstanceCountInResource') is not None:
+            temp_model = ServiceInstanceCountInResource()
+            self.instance_count_in_resource = temp_model.from_map(m['InstanceCountInResource'])
         if m.get('InternetEndpoint') is not None:
             self.internet_endpoint = m.get('InternetEndpoint')
         if m.get('IntranetEndpoint') is not None:
@@ -1052,6 +1176,8 @@ class Service(TeaModel):
             self.resource = m.get('Resource')
         if m.get('ResourceAlias') is not None:
             self.resource_alias = m.get('ResourceAlias')
+        if m.get('ResourceBurstable') is not None:
+            self.resource_burstable = m.get('ResourceBurstable')
         if m.get('Role') is not None:
             self.role = m.get('Role')
         if m.get('RoleAttrs') is not None:
@@ -7238,9 +7364,13 @@ class DescribeResourceResponseBody(TeaModel):
         cpu_used: int = None,
         create_time: str = None,
         extra_data: str = None,
+        features: List[str] = None,
         gpu_count: int = None,
         gpu_used: float = None,
         instance_count: int = None,
+        instance_max_allocatable_cpu: int = None,
+        instance_max_allocatable_gpu: float = None,
+        instance_max_allocatable_memory: int = None,
         memory: int = None,
         memory_used: int = None,
         message: str = None,
@@ -7264,12 +7394,16 @@ class DescribeResourceResponseBody(TeaModel):
         self.create_time = create_time
         # The additional information, such as the connection status of a virtual private cloud (VPC) and the log status of Log Service.
         self.extra_data = extra_data
+        self.features = features
         # The total number of GPUs.
         self.gpu_count = gpu_count
         # The number of GPUs that is used.
         self.gpu_used = gpu_used
         # The total number of instances in the resource group.
         self.instance_count = instance_count
+        self.instance_max_allocatable_cpu = instance_max_allocatable_cpu
+        self.instance_max_allocatable_gpu = instance_max_allocatable_gpu
+        self.instance_max_allocatable_memory = instance_max_allocatable_memory
         # The total memory size. Unit: MB.
         self.memory = memory
         # The size of memory that is used. Unit: MB.
@@ -7317,12 +7451,20 @@ class DescribeResourceResponseBody(TeaModel):
             result['CreateTime'] = self.create_time
         if self.extra_data is not None:
             result['ExtraData'] = self.extra_data
+        if self.features is not None:
+            result['Features'] = self.features
         if self.gpu_count is not None:
             result['GpuCount'] = self.gpu_count
         if self.gpu_used is not None:
             result['GpuUsed'] = self.gpu_used
         if self.instance_count is not None:
             result['InstanceCount'] = self.instance_count
+        if self.instance_max_allocatable_cpu is not None:
+            result['InstanceMaxAllocatableCPU'] = self.instance_max_allocatable_cpu
+        if self.instance_max_allocatable_gpu is not None:
+            result['InstanceMaxAllocatableGPU'] = self.instance_max_allocatable_gpu
+        if self.instance_max_allocatable_memory is not None:
+            result['InstanceMaxAllocatableMemory'] = self.instance_max_allocatable_memory
         if self.memory is not None:
             result['Memory'] = self.memory
         if self.memory_used is not None:
@@ -7361,12 +7503,20 @@ class DescribeResourceResponseBody(TeaModel):
             self.create_time = m.get('CreateTime')
         if m.get('ExtraData') is not None:
             self.extra_data = m.get('ExtraData')
+        if m.get('Features') is not None:
+            self.features = m.get('Features')
         if m.get('GpuCount') is not None:
             self.gpu_count = m.get('GpuCount')
         if m.get('GpuUsed') is not None:
             self.gpu_used = m.get('GpuUsed')
         if m.get('InstanceCount') is not None:
             self.instance_count = m.get('InstanceCount')
+        if m.get('InstanceMaxAllocatableCPU') is not None:
+            self.instance_max_allocatable_cpu = m.get('InstanceMaxAllocatableCPU')
+        if m.get('InstanceMaxAllocatableGPU') is not None:
+            self.instance_max_allocatable_gpu = m.get('InstanceMaxAllocatableGPU')
+        if m.get('InstanceMaxAllocatableMemory') is not None:
+            self.instance_max_allocatable_memory = m.get('InstanceMaxAllocatableMemory')
         if m.get('Memory') is not None:
             self.memory = m.get('Memory')
         if m.get('MemoryUsed') is not None:
@@ -9340,6 +9490,7 @@ class DescribeVirtualResourceResponseBody(TeaModel):
         self,
         create_time: str = None,
         disable_spot_protection_period: bool = None,
+        features: List[str] = None,
         request_id: str = None,
         resources: List[DescribeVirtualResourceResponseBodyResources] = None,
         service_count: int = None,
@@ -9351,6 +9502,7 @@ class DescribeVirtualResourceResponseBody(TeaModel):
         self.create_time = create_time
         # Indicates whether the retention period of preemptible instances was disabled.
         self.disable_spot_protection_period = disable_spot_protection_period
+        self.features = features
         # The ID of the request.
         self.request_id = request_id
         # The list of resources in the virtual resource group.
@@ -9380,6 +9532,8 @@ class DescribeVirtualResourceResponseBody(TeaModel):
             result['CreateTime'] = self.create_time
         if self.disable_spot_protection_period is not None:
             result['DisableSpotProtectionPeriod'] = self.disable_spot_protection_period
+        if self.features is not None:
+            result['Features'] = self.features
         if self.request_id is not None:
             result['RequestId'] = self.request_id
         result['Resources'] = []
@@ -9402,6 +9556,8 @@ class DescribeVirtualResourceResponseBody(TeaModel):
             self.create_time = m.get('CreateTime')
         if m.get('DisableSpotProtectionPeriod') is not None:
             self.disable_spot_protection_period = m.get('DisableSpotProtectionPeriod')
+        if m.get('Features') is not None:
+            self.features = m.get('Features')
         if m.get('RequestId') is not None:
             self.request_id = m.get('RequestId')
         self.resources = []
@@ -13213,6 +13369,8 @@ class ListServiceVersionsResponse(TeaModel):
 class ListServicesRequest(TeaModel):
     def __init__(
         self,
+        autoscaler_enabled: bool = None,
+        cronscaler_enabled: bool = None,
         filter: str = None,
         gateway: str = None,
         group_name: str = None,
@@ -13224,6 +13382,7 @@ class ListServicesRequest(TeaModel):
         parent_service_uid: str = None,
         quota_id: str = None,
         resource_alias_name: str = None,
+        resource_burstable: bool = None,
         resource_id: str = None,
         resource_name: str = None,
         resource_type: str = None,
@@ -13235,6 +13394,8 @@ class ListServicesRequest(TeaModel):
         sort: str = None,
         workspace_id: str = None,
     ):
+        self.autoscaler_enabled = autoscaler_enabled
+        self.cronscaler_enabled = cronscaler_enabled
         # The field that is used for fuzzy matches. The system performs fuzzy matches only by service name.
         self.filter = filter
         # The private gateway ID.
@@ -13258,6 +13419,7 @@ class ListServicesRequest(TeaModel):
         # The quota ID.
         self.quota_id = quota_id
         self.resource_alias_name = resource_alias_name
+        self.resource_burstable = resource_burstable
         self.resource_id = resource_id
         # The name or ID of the resource group to which the service belongs.
         self.resource_name = resource_name
@@ -13483,6 +13645,10 @@ class ListServicesRequest(TeaModel):
             return _map
 
         result = dict()
+        if self.autoscaler_enabled is not None:
+            result['AutoscalerEnabled'] = self.autoscaler_enabled
+        if self.cronscaler_enabled is not None:
+            result['CronscalerEnabled'] = self.cronscaler_enabled
         if self.filter is not None:
             result['Filter'] = self.filter
         if self.gateway is not None:
@@ -13505,6 +13671,8 @@ class ListServicesRequest(TeaModel):
             result['QuotaId'] = self.quota_id
         if self.resource_alias_name is not None:
             result['ResourceAliasName'] = self.resource_alias_name
+        if self.resource_burstable is not None:
+            result['ResourceBurstable'] = self.resource_burstable
         if self.resource_id is not None:
             result['ResourceId'] = self.resource_id
         if self.resource_name is not None:
@@ -13529,6 +13697,10 @@ class ListServicesRequest(TeaModel):
 
     def from_map(self, m: dict = None):
         m = m or dict()
+        if m.get('AutoscalerEnabled') is not None:
+            self.autoscaler_enabled = m.get('AutoscalerEnabled')
+        if m.get('CronscalerEnabled') is not None:
+            self.cronscaler_enabled = m.get('CronscalerEnabled')
         if m.get('Filter') is not None:
             self.filter = m.get('Filter')
         if m.get('Gateway') is not None:
@@ -13551,6 +13723,8 @@ class ListServicesRequest(TeaModel):
             self.quota_id = m.get('QuotaId')
         if m.get('ResourceAliasName') is not None:
             self.resource_alias_name = m.get('ResourceAliasName')
+        if m.get('ResourceBurstable') is not None:
+            self.resource_burstable = m.get('ResourceBurstable')
         if m.get('ResourceId') is not None:
             self.resource_id = m.get('ResourceId')
         if m.get('ResourceName') is not None:
@@ -13577,6 +13751,8 @@ class ListServicesRequest(TeaModel):
 class ListServicesShrinkRequest(TeaModel):
     def __init__(
         self,
+        autoscaler_enabled: bool = None,
+        cronscaler_enabled: bool = None,
         filter: str = None,
         gateway: str = None,
         group_name: str = None,
@@ -13588,6 +13764,7 @@ class ListServicesShrinkRequest(TeaModel):
         parent_service_uid: str = None,
         quota_id: str = None,
         resource_alias_name: str = None,
+        resource_burstable: bool = None,
         resource_id: str = None,
         resource_name: str = None,
         resource_type: str = None,
@@ -13599,6 +13776,8 @@ class ListServicesShrinkRequest(TeaModel):
         sort: str = None,
         workspace_id: str = None,
     ):
+        self.autoscaler_enabled = autoscaler_enabled
+        self.cronscaler_enabled = cronscaler_enabled
         # The field that is used for fuzzy matches. The system performs fuzzy matches only by service name.
         self.filter = filter
         # The private gateway ID.
@@ -13622,6 +13801,7 @@ class ListServicesShrinkRequest(TeaModel):
         # The quota ID.
         self.quota_id = quota_id
         self.resource_alias_name = resource_alias_name
+        self.resource_burstable = resource_burstable
         self.resource_id = resource_id
         # The name or ID of the resource group to which the service belongs.
         self.resource_name = resource_name
@@ -13847,6 +14027,10 @@ class ListServicesShrinkRequest(TeaModel):
             return _map
 
         result = dict()
+        if self.autoscaler_enabled is not None:
+            result['AutoscalerEnabled'] = self.autoscaler_enabled
+        if self.cronscaler_enabled is not None:
+            result['CronscalerEnabled'] = self.cronscaler_enabled
         if self.filter is not None:
             result['Filter'] = self.filter
         if self.gateway is not None:
@@ -13869,6 +14053,8 @@ class ListServicesShrinkRequest(TeaModel):
             result['QuotaId'] = self.quota_id
         if self.resource_alias_name is not None:
             result['ResourceAliasName'] = self.resource_alias_name
+        if self.resource_burstable is not None:
+            result['ResourceBurstable'] = self.resource_burstable
         if self.resource_id is not None:
             result['ResourceId'] = self.resource_id
         if self.resource_name is not None:
@@ -13893,6 +14079,10 @@ class ListServicesShrinkRequest(TeaModel):
 
     def from_map(self, m: dict = None):
         m = m or dict()
+        if m.get('AutoscalerEnabled') is not None:
+            self.autoscaler_enabled = m.get('AutoscalerEnabled')
+        if m.get('CronscalerEnabled') is not None:
+            self.cronscaler_enabled = m.get('CronscalerEnabled')
         if m.get('Filter') is not None:
             self.filter = m.get('Filter')
         if m.get('Gateway') is not None:
@@ -13915,6 +14105,8 @@ class ListServicesShrinkRequest(TeaModel):
             self.quota_id = m.get('QuotaId')
         if m.get('ResourceAliasName') is not None:
             self.resource_alias_name = m.get('ResourceAliasName')
+        if m.get('ResourceBurstable') is not None:
+            self.resource_burstable = m.get('ResourceBurstable')
         if m.get('ResourceId') is not None:
             self.resource_id = m.get('ResourceId')
         if m.get('ResourceName') is not None: